Move Brooklyn classes to org.apache
