+++ clocker.bom
@@ -97,17 +97,17 @@ brooklyn.catalog:
   - id: etcd-node
     name: "Etcd Node"
     item:
-      type: brooklyn.entity.nosql.etcd.EtcdNode
+      type: org.apache.brooklyn.entity.nosql.etcd.EtcdNode
 
   - id: etcd-proxy
     name: "Etcd Proxy"
     item:
-      type: brooklyn.entity.nosql.etcd.EtcdProxy
+      type: org.apache.brooklyn.entity.nosql.etcd.EtcdProxy
 
   - id: etcd-cluster
     name: "Etcd Cluster"
     item:
-      type: brooklyn.entity.nosql.etcd.EtcdCluster
+      type: org.apache.brooklyn.entity.nosql.etcd.EtcdCluster
       brooklyn.config:
         etcd.version: 2.0.11
         etcd.node.spec:
@@ -231,5 +231,24 @@ brooklyn.catalog:
       - type: mesos-cluster
         brooklyn.config:
           mesos.url: "http://localhost:5050/"
+
+  - id: haproxy-controller
+    name: "HAProxy"
+    description: |
+      HAProxy load balancer entity.
+    itemType: template
+    type: org.apache.brooklyn.entity.proxy.haproxy.HAProxyController
+    brooklyn.config:
+      docker.image.name: haproxy
+      docker.image.tag: 1.5.9
+      docker.container.directPorts.configKeys:
+      - $brooklyn:sensor("org.apache.brooklyn.entity.proxy.AbstractController", "proxy.http.port")
+      - $brooklyn:sensor("org.apache.brooklyn.entity.proxy.AbstractController", "proxy.https.port")
+      install.dir: /usr/local/sbin/
+      run.dir: /usr/local/etc/haproxy/
+      member.sensor.hostname: "host.subnet.hostname"
+      provisioning.properties:
+        entity.dynamicLocation.flags:
+          user: root
 
 # vim:ts=2:sw=2:ft=yaml:
