+++ docker/src/main/java/clocker/docker/networking/entity/sdn/calico/CalicoNode.java
@@ -21,10 +21,9 @@ import org.apache.brooklyn.api.entity.ImplementedBy;
 import org.apache.brooklyn.config.ConfigKey;
 import org.apache.brooklyn.core.config.ConfigKeys;
 import org.apache.brooklyn.core.sensor.AttributeSensorAndConfigKey;
+import org.apache.brooklyn.entity.nosql.etcd.EtcdNode;
 import org.apache.brooklyn.entity.software.base.SoftwareProcess;
 import org.apache.brooklyn.util.core.flags.SetFromFlag;
-
-import brooklyn.entity.nosql.etcd.EtcdNode;
 
 /**
  * The Calico plugin
