+++ extras/src/main/java/org/apache/brooklyn/entity/proxy/haproxy/HAProxyDriver.java
@@ -13,9 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package brooklyn.entity.nosql.etcd;
+package org.apache.brooklyn.entity.proxy.haproxy;
 
+import org.apache.brooklyn.entity.software.base.SoftwareProcessDriver;
 
-public interface EtcdProxyDriver extends EtcdNodeDriver {
+public interface HAProxyDriver extends SoftwareProcessDriver {
+
+    void reconfigureService();
 
 }
