+++ mesos/src/main/java/clocker/mesos/networking/entity/sdn/calico/CalicoModule.java
@@ -27,8 +27,7 @@ import org.apache.brooklyn.api.entity.ImplementedBy;
 import org.apache.brooklyn.config.ConfigKey;
 import org.apache.brooklyn.core.config.ConfigKeys;
 import org.apache.brooklyn.core.sensor.AttributeSensorAndConfigKey;
-
-import brooklyn.entity.nosql.etcd.EtcdProxy;
+import org.apache.brooklyn.entity.nosql.etcd.EtcdProxy;
 
 /**
  * The Calico NetworkModule for Mesos.
